simlified the qualifier summary by adding to SearchResultSummary
